[fix][client] Fix internal receive used wrong timeout type (#15014)
### Motivation

Currently, when we called `redeliverUnacknowledgedMessages` and consumer use receive with timeout unit is second, the timeout might not work.

When we use the second as receive timeout unit like `consumer.receive(3, TimeUnit.SECONDS);`, line 473 will pass a negative timeout to the next `internalReceive` calls, because we are subtracting nanoseconds from seconds.

https://github.com/apache/pulsar/blob/f3b87b65c6946eb197c1eece22cff8ff04e16fcb/pulsar-client/src/main/java/org/apache/pulsar/client/impl/ConsumerImpl.java#L459-L476

### Modifications

* Use the same type to subtract.
* Use long type as timeout argument.

(cherry picked from commit bdc302426dba955652c743e20eef853c34849a24)

+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/ConsumerBase.java
@@ -198,7 +198,7 @@ public abstract class ConsumerBase<T> extends HandlerState implements Consumer<T
         return internalReceive(timeout, unit);
     }
 
-    protected abstract Message<T> internalReceive(int timeout, TimeUnit unit) throws PulsarClientException;
+    protected abstract Message<T> internalReceive(long timeout, TimeUnit unit) throws PulsarClientException;
 
     @Override
     public Messages<T> batchReceive() throws PulsarClientException {
